// app-eurofish-legacy/src/app/notificacion/interfaces/eufi.ts
export interface IResponseEUFI {
  'odata.metadata': string;
  AbsoluteEntry: number;
  DocumentNumber: number;
  Series: number;
  ItemNo: string;
  ProductionOrderStatus: string;
  ProductionOrderType: string;
  PlannedQuantity: number;
  CompletedQuantity: number;
  RejectedQuantity: number;
  PostingDate: string;
  DueDate: string;
  ProductionOrderOriginEntry: null;
  ProductionOrderOriginNumber: null;
  ProductionOrderOrigin: string;
  UserSignature: number;
  Remarks: null;
  ClosingDate: null;
  ReleaseDate: string;
  CustomerCode: null;
  Warehouse: string;
  InventoryUOM: string;
  JournalRemarks: string;
  TransactionNumber: null;
  CreationDate: string;
  Printed: string;
  DistributionRule: string;
  Project: null;
  DistributionRule2: string;
  DistributionRule3: string;
  DistributionRule4: string;
  DistributionRule5: string;
  UoMEntry: number;
  StartDate: string;
  ProductDescription: string;
  Priority: number;
  RoutingDateCalculation: string;
  UpdateAllocation: string;
  SAPPassport: null;
  AttachmentEntry: null;
  U_SYP_ORDEN_PADRE: null;
  U_SYP_ORDEN_N1: null;
  U_SYP_ORDEN_N2: null;
  U_SYP_ORDEN_N3: null;
  ProductionOrderLines: ProductionOrderLine[];
  ProductionOrdersSalesOrderLines: any[];
  ProductionOrdersStages: any[];
  ProductionOrdersDocumentReferences: any[];
}

export interface ProductionOrderLine {
  DocumentAbsoluteEntry: number;
  LineNumber: number;
  ItemNo: string;
  BaseQuantity: number;
  PlannedQuantity: number;
  IssuedQuantity: number;
  ProductionOrderIssueType: string;
  Warehouse: string;
  VisualOrder: number;
  DistributionRule: null;
  LocationCode: null;
  Project: null;
  DistributionRule2: null;
  DistributionRule3: null;
  DistributionRule4: null;
  DistributionRule5: null;
  UoMEntry: number;
  UoMCode: number;
  WipAccount: string;
  ItemType: string;
  LineText: null;
  AdditionalQuantity: number;
  ResourceAllocation: null | string;
  StartDate: string;
  EndDate: string;
  StageID: null;
  RequiredDays: number;
  ItemName: string;
  WeightOfRecycledPlastic: number;
  PlasticPackageExemptionReason: null;
  SerialNumbers: any[];
  BatchNumbers: any[];
  isNotStockItem?: number;
}